requirements.txt
amadeus==12.0.0
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.4
celery==5.5.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Deprecated==1.3.1
dj-database-url==1.0.0
Django==4.2.28
django-celery-beat==2.9.0
django-cors-headers==4.4.0
django-countries==8.2.0
django-prometheus==2.4.1
django-redis==5.4.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.5.0
inflection==0.5.1
kombu==5.5.4
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==26.0
pillow==10.4.0
pkg_resources==0.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==6.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.13.2
tzdata==2025.3
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.7.0
wrapt==2.0.1
zipp==3.20.2
